Chnages to Home table and style.css
A bootstrap table template was used to constrict the table that appears on our homepage which has the ability to sort the entries that are listed, search the whole table, and chosse how many entries are seen at one time. Also on the table an additional field was added so that the needs that each client has is shown as well.